+++ src/main/java/dinkplugin/domain/PlayerLookupService.java
@@ -7,8 +7,7 @@ public enum PlayerLookupService {
     OSRS_HISCORE("OSRS HiScore"),
     CRYSTAL_MATH_LABS("Crystal Math Labs"),
     TEMPLE_OSRS("Temple OSRS"),
-    WISE_OLD_MAN("Wise Old Man"),
-    COLLECTION_LOG("CollectionLog.net");
+    WISE_OLD_MAN("Wise Old Man");
 
     private final String name;
 
@@ -27,8 +26,6 @@ public enum PlayerLookupService {
                 return "https://crystalmathlabs.com/track.php?player=" + escapedName;
             case TEMPLE_OSRS:
                 return "https://templeosrs.com/player/overview.php?player=" + escapedName;
-            case COLLECTION_LOG:
-                return "https://collectionlog.net/log/" + escapedName;
             case NONE:
             default:
                 return null;